Move 3D panel settings sync buttons to topic picker, and other settings to sidebar
