+++ app/panels/ThreeDimensionalViz/TopicTree/Layout.tsx
@@ -38,6 +38,7 @@ import useDataSourceInfo from "@foxglove-studio/app/PanelAPI/useDataSourceInfo";
 import KeyListener from "@foxglove-studio/app/components/KeyListener";
 import Modal from "@foxglove-studio/app/components/Modal";
 import PanelContext from "@foxglove-studio/app/components/PanelContext";
+import PanelToolbar from "@foxglove-studio/app/components/PanelToolbar";
 import { RenderToBodyComponent } from "@foxglove-studio/app/components/RenderToBodyComponent";
 import useGlobalVariables from "@foxglove-studio/app/hooks/useGlobalVariables";
 import useShallowMemo from "@foxglove-studio/app/hooks/useShallowMemo";
@@ -789,8 +790,6 @@ export default function Layout({
   ]);
 
   const cursorType = isDrawing ? "crosshair" : "";
-  const { isHovered = false } = useContext(PanelContext) ?? {};
-  const isHidden = !isHovered;
 
   const { videoRecordingStyle } = useMemo(
     () => ({
@@ -831,14 +830,7 @@ export default function Layout({
           data-test="3dviz-layout"
         >
           <KeyListener keyDownHandlers={keyDownHandlers} />
-          <PanelToolbarMenu
-            autoTextBackgroundColor={autoTextBackgroundColor}
-            checkedKeys={checkedKeys}
-            flattenMarkers={!!flattenMarkers}
-            helpContent={helpContent}
-            saveConfig={saveConfig}
-            settingsByKey={settingsByKey}
-          />
+          <PanelToolbar floating helpContent={helpContent} />
           <div style={{ position: "absolute", width: "100%", height: "100%" }}>
             <div
               style={
@@ -851,37 +843,36 @@ export default function Layout({
               }
               ref={topicTreeSizeRef}
             >
-              {isHovered && (
-                <TopicTree
-                  allKeys={allKeys}
-                  availableNamespacesByTopic={availableNamespacesByTopic}
-                  checkedKeys={checkedKeys}
-                  containerHeight={containerHeight ?? 0}
-                  containerWidth={containerWidth ?? 0}
-                  derivedCustomSettingsByKey={derivedCustomSettingsByKey}
-                  expandedKeys={expandedKeys}
-                  filterText={filterText}
-                  getIsNamespaceCheckedByDefault={getIsNamespaceCheckedByDefault}
-                  getIsTreeNodeVisibleInScene={getIsTreeNodeVisibleInScene}
-                  getIsTreeNodeVisibleInTree={getIsTreeNodeVisibleInTree}
-                  hasFeatureColumn={hasFeatureColumn}
-                  onExitTopicTreeFocus={onExitTopicTreeFocus}
-                  onNamespaceOverrideColorChange={onNamespaceOverrideColorChange}
-                  pinTopics={pinTopics}
-                  diffModeEnabled={diffModeEnabled}
-                  rootTreeNode={rootTreeNode}
-                  saveConfig={saveConfig}
-                  sceneErrorsByKey={sceneErrorsByKey}
-                  setCurrentEditingTopic={setCurrentEditingTopic}
-                  setEditingNamespace={setEditingNamespace}
-                  setFilterText={setFilterText}
-                  setShowTopicTree={setShowTopicTree}
-                  shouldExpandAllKeys={shouldExpandAllKeys}
-                  showTopicTree={showTopicTree}
-                  topicDisplayMode={topicDisplayMode}
-                  visibleTopicsCountByKey={visibleTopicsCountByKey}
-                />
-              )}
+              <TopicTree
+                allKeys={allKeys}
+                availableNamespacesByTopic={availableNamespacesByTopic}
+                checkedKeys={checkedKeys}
+                containerHeight={containerHeight ?? 0}
+                containerWidth={containerWidth ?? 0}
+                settingsByKey={settingsByKey}
+                derivedCustomSettingsByKey={derivedCustomSettingsByKey}
+                expandedKeys={expandedKeys}
+                filterText={filterText}
+                getIsNamespaceCheckedByDefault={getIsNamespaceCheckedByDefault}
+                getIsTreeNodeVisibleInScene={getIsTreeNodeVisibleInScene}
+                getIsTreeNodeVisibleInTree={getIsTreeNodeVisibleInTree}
+                hasFeatureColumn={hasFeatureColumn}
+                onExitTopicTreeFocus={onExitTopicTreeFocus}
+                onNamespaceOverrideColorChange={onNamespaceOverrideColorChange}
+                pinTopics={pinTopics}
+                diffModeEnabled={diffModeEnabled}
+                rootTreeNode={rootTreeNode}
+                saveConfig={saveConfig}
+                sceneErrorsByKey={sceneErrorsByKey}
+                setCurrentEditingTopic={setCurrentEditingTopic}
+                setEditingNamespace={setEditingNamespace}
+                setFilterText={setFilterText}
+                setShowTopicTree={setShowTopicTree}
+                shouldExpandAllKeys={shouldExpandAllKeys}
+                showTopicTree={showTopicTree}
+                topicDisplayMode={topicDisplayMode}
+                visibleTopicsCountByKey={visibleTopicsCountByKey}
+              />
               {currentEditingTopic && (
                 <TopicSettingsModal
                   currentEditingTopic={currentEditingTopic}
@@ -970,7 +961,6 @@ export default function Layout({
                   targetPose={targetPose}
                   transforms={transforms}
                   rootTf={rootTf}
-                  isHidden={isHidden}
                   {...searchTextProps}
                 />
               </div>
